requirements.txt
#
# This file is autogenerated by pip-compile
# To update, run:
#
#    pip-compile requirements.in
#
--extra-index-url https://www.piwheels.org/simple

aiofiles==0.5.0           # via sanic
certifi==2020.4.5.1       # via httpx, requests
chardet==3.0.4            # via httpx, requests
click==7.1.1              # via sonos-cli
h11==0.8.1                # via httpx
h2==3.2.0                 # via httpx
hpack==3.0.0              # via h2
hstspreload==2020.4.14    # via httpx
httptools==0.1.1          # via sanic
httpx==0.9.3              # via sanic
hyperframe==5.2.0         # via h2
idna==2.9                 # via httpx, requests
mfrc522==0.0.7            # via -r requirements.in
multidict==4.7.5          # via sanic
oauthlib==3.1.0           # via requests-oauthlib
requests-oauthlib==1.3.0  # via sonos-cli
requests==2.23.0          # via requests-oauthlib, sonos-cli
rfc3986==1.4.0            # via httpx
rpi.gpio==0.7.0           # via mfrc522
sanic==19.12.2            # via -r requirements.in
sniffio==1.1.0            # via httpx
sonos-cli==0.3.5          # via -r requirements.in
spidev==3.4               # via -r requirements.in, mfrc522
tabulate==0.8.7           # via sonos-cli
ujson==2.0.3              # via sanic
urllib3==1.25.9           # via requests
uvloop==0.14.0            # via sanic
websockets==8.1           # via sanic
